requirements.txt
amqp==5.1.0
anyascii==0.3.1
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
captcha==0.3
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.11.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==21.1
django-modelcluster==5.3
django-nine==0.2.5
django-rest-captcha==0.2.0
django-rest-framework==0.1.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
draftjs-exporter==2.1.7
drf-spectacular==0.22.1
elasticsearch==7.17.2
elasticsearch-django==7.1.3
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
executing==0.8.3
glob2==0.7
html5lib==1.1
icecream==2.1.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
jalalidate==0.1.2
Jinja2==3.1.2
jsonschema==4.5.1
kombu==5.2.4
l18n==2021.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.3
# openapi-codec==1.3.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
openpyxl==3.0.9
persiantools==3.0.0
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylibjpeg-libjpeg==1.3.1
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
toml==0.10.2
tzdata==2021.5
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
vine==5.0.0
wagtail==2.16.2
wcwidth==0.2.5
webencodings==0.5.1
wheezy.captcha==3.0.2
Willow==1.4.1
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zlib-state==0.1.5
